Ensure LaTeX equation environments are processed by pandoc.
Works around a bug in pandoc (https://github.com/jgm/pandoc/issues/1764,
present in at least up-to and including pandoc 1.18) to surround every
AMSMath latex environment in the org file with plain TeX equation block
delimiters, '$$ .. $$' in order for pandoc to properly recognise the
maths environment as a latex equation.  Also adds surrounding
line-breaks so that pandoc treats the math environment as its own
paragraph.  This avoids having text before or after the math
environment ending up on the same line as the equation.