Number `Table' captions in pandoc export
Works around a bug in pandoc (present in at least up to and including
pandoc 1.18), where captions for Tables/Figures/Equations/etc. are not
numbered, unlike in other org-mode exporters.

This adds a function `org-pandoc-set-caption-title' to add a numbered
label (`Table %d', `Figure %d' etc.) to the caption.

Additionally, pandoc doesn't seem to pick up on #+LABEL: specifiers for
link targets to Tables/Figures/etc., so this function additionally adds
a manual target for any hyperlinks.

Overrides the table export function to add this caption processing
before sending the table to pandoc.