BUGFIX: handle the case when captions start with non-plain text
If the first `element' in a caption is, e.g., a latex-fragment, or bold,
or italic or similar, then simply setting the `cl-caaar' of the caption
to pre-pend a target for pandoc links doesn't work.

Instead, it's more robust to insert an entirely new element into the
`caar' of the caption that just contains the appropriate link target,
and reference name if needed.